requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
psycopg2-binary
python-dotenv